+++ changelog.md
@@ -25,3 +25,4 @@ Elisif-Simple was created by: **Cannicide#2753**
 - Added ability to use semi-HTML syntax in sent messages and replies, for markdown, embeds, attachments, and message components
     - This feature is really cool. Advanced features like embeds and buttons can be added to a message, all within the content String.
     - Markup buttons now support built-in, functional "onclick" handlers! Abstracts away tons of ugly discord.js component collector code.
+    - Replies can now be made ephemeral via markup.